third_party/blink/web_tests/fast/events/pointerevents/mouse-pointer-on-scrollbar.html

<!DOCTYPE HTML>
<script src="../../../resources/js-test.js"></script>
<style>
#target {
  height: 100px;
  width: 100px;
  overflow-y: scroll;
}
#spacer {
  background: green;
  height: 200px;
}
</style>

<div id="target">
<div id="spacer"></div>
</div>

<div id="console"></div>

<script>
description("Verifies that pointerup/down are fired correctly for corresponding mouse events on the scollbar.");
var targetDiv = document.getElementById('target');

function init() {
  var eventList = ["mousedown", "mouseup", "pointerdown", "pointerup"];
  eventList.forEach(function(eventName) {
    targetDiv.addEventListener(eventName, function(event) {
      debug("Received " + event.type);
    });
  });
}

function runTests() {
  var rect = targetDiv.getBoundingClientRect();
  var x1 = rect.right - 5;
  var y1 = rect.top + 20;
  if (targetDiv != document.elementFromPoint(x1,y1))
    testFailed("Didn't hit the scrollbar as expected");
  eventSender.mouseMoveTo(x1, y1);
  eventSender.mouseDown();
  eventSender.mouseUp();
}

init();
if (window.eventSender)
  runTests();
else
  debug("This test requires eventSender");

</script>
