third_party/WebKit/LayoutTests/fast/events/pointerevents/mouse-pointer-on-scrollbar-expected.txt

Verifies that pointerup/down are fired correctly for corresponding mouse events on the scollbar.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Received pointerdown
Received mousedown
Received pointerup
Received mouseup
PASS successfullyParsed is true

TEST COMPLETE
